manejo de null en el uno a muchos creación de tipo de producto por defecto dentro del pojo de producto

+++ src/main/java/net/daw/operation/ProductoRellena1.java
@@ -74,7 +74,7 @@ public class ProductoRellena1 extends Operation {
                     oProductoBean.setId(0);
                     oProductoBean.setCodigo(Character.toUpperCase(primero.charAt(0)) + Character.toUpperCase(segundo.charAt(0)) + contador.toString() + Character.toUpperCase(tercero.charAt(0)));
                     oProductoBean.setDescripcion(primero + " " + segundo + " " + tercero);
-                    oProductoBean.setPrecio(Float.parseFloat(Integer.toString(generator.nextInt(2000)) + "." + Integer.toString(generator.nextInt(99))));
+                    oProductoBean.setPrecio(Double.parseDouble(Integer.toString(generator.nextInt(2000)) + "." + Integer.toString(generator.nextInt(99))));
                     try {
                         oProductoDao.set(oProductoBean);
